graph_bfs is an algorithm for breadth first search
The program shall read in a graph in adjacency list format and perform a breadth first search (BFS) on the graph, starting at the node first read. Each node's name is a single letter 'A'-'Z', so you can safely assume less than 27 nodes for the graph. The format of each line to be read in is specified as
  <nodeName>-<string of neighboring node names>\n   (see example below).
The requested output  of the program is one line of node names followed by '\n', in which the order of names represents one possible exploration following breadth first search.
Note (1): for most inputs several different outputs are possible. Here we only ask you to find one of them!
Note (2): all graphs here are undirected, i.e. if a connection A-...B... exists you will also find a connection B-...A....