this algorithm searches depth first
the program shall read in a graph in adjacency list format and perform a Depth-First-Search (DFS) on the graph, starting at the node first read. Each node's name is a single letter 'A'-'Z', so you can safely assume less than 27 nodes for the graph. The format of each line to be read in is specified as
  <nodeName>-<string of neighboring-node-names>\n  
The requested output  of the program is one line of node names followed by '\n', in which the order of names represents one possible exploration following depth first search.
Note (1): for most inputs several different outputs are possible. 
Note (2): all graphs here are undirected, i.e. if a connection A-...B... exists your will also find a connection B-...A....