implementation of dijkstra algorithm
the program shall read in a graph with associated costs for traversing edges, and find optimal paths from a starting node to all other nodes in this graph. This is the same problem as finding shortest distances on highways that connect cities! Each node's name is a single letter 'A'-'Z', so you can safely assume less than 27 nodes for the graph. Each connected pair of nodes will be provided as an independent input line in the following format:
  <nodeNameA>-<nodeNameB>-<cost>\n   (see example below).
You can assume the cost of traveling between neighboring nodes to be a positive integer number below 100 for each single edge: 0 < edge-cost < 100.
the program needs to traverse the graph following Dijkstra's algorithm to find the shortest routes from a given starting node S to all other nodes in the graph. S is the one node that was mentioned first in the first line of input. When done, for each node N that exists in the graph, the program shall provide the backtracking p